requirements.txt
Django>=5.0,<6
djangorestframework>=3.15
pydantic>=2.0
httpx>=0.27
crewai>=0.95
json5>=0.9.0
Pillow>=10.0
pymupdf>=1.24.0
openpyxl>=3.1.0
whitenoise>=6.6
python-dotenv>=1.0

# dev / test
pytest>=8.0
pytest-django>=4.8
factory-boy>=3.3
